requirements.txt
# DeerFlow Hello World Demo - Python Dependencies
# Install with: pip install -r requirements.txt
# Or let uv handle it: uv sync (recommended)

# =============================================================================
# CORE DEPENDENCIES (Required)
# =============================================================================

# Environment variable management
python-dotenv>=1.0.0

# HTTP requests library
requests>=2.31.0

# =============================================================================
# DEERFLOW CORE (Installed via uv sync from deer-flow repository)
# The following are typically installed by DeerFlow itself:
# =============================================================================

# DeerFlow client library - provides DeerFlowClient
# Install from local source after running 'uv sync' in deer-flow/backend
# Uncomment the line below and adjust path as needed:
# -e ./deer-flow/backend/packages/harness

# Alternative: Install via uv workspace (recommended)
# cd deer-flow/backend && uv sync

# LangChain framework
# langchain>=0.3.0
# langchain-core>=0.3.0
# langchain-openai>=0.2.0
# langchain-anthropic>=0.2.0
# langgraph>=0.2.0

# Async support
# httpx>=0.26.0
# aiohttp>=3.9.0

# Configuration management
# pyyaml>=6.0
# pydantic>=2.5.0
# pydantic-settings>=2.1.0

# =============================================================================
# OPTIONAL: For standalone hello_world.py script
# (These are needed if running hello_world.py outside DeerFlow)
# =============================================================================

# For checking prerequisites
psutil>=5.9.0

# =============================================================================
# DEVELOPMENT TOOLS (Optional, for debugging)
# =============================================================================

# Logging enhancement
# colorama>=0.4.6

# Testing (if you want to write tests)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0

# =============================================================================
# NOTES:
# 1. Main dependencies are managed by DeerFlow in deer-flow/pyproject.toml
# 2. Run 'uv sync' in the deer-flow directory to install all dependencies
# 3. This file is primarily for the hello_world.py demo script
# 4. Core requirements: python-dotenv, requests, psutil
# =============================================================================
